requirements.txt
asgiref==3.11.0
Django==6.0
sqlparse==0.5.4
tzdata==2025.2
cryptography==43.0.0
abacatepay>=0.1.0
python-decouple>=3.6
requests>=2.31.0
